Try another approach to make File.Select more reliable
Last time, I reused the same <input> node again and again. This
resolved the GC problem, but one aspect of file inputs is that if you
select the same files, no event is triggered when you press "Ok" on the
dialog. It did not "change".

This time, I am keeping the input nodes alive with a global variable,
but allocating a new <input> on each command. I believe this will
resolve the GC flakiness and resolve the "what if someone selects the
same thing" problem.