Make the File.Download.url warning more accurate
It appears that things work on IE, but in a degraded way, as if it was
a cross-origin request.

It may be necessary to add the check for the download attribute to add
(node.target = "_blank") in this case, but the current code works
alright with IE11 as is. Not sure about IE10.